requirements.txt
google-adk>=1.0.0
google-cloud-aiplatform[adk,agent_engines]>=1.108.0
pyyaml>=6.0
python-dotenv>=1.0.0
pydantic>=2.11.7
cloudpickle>=3.0.0